Initial checkin
This adds a good amount of code, suitable for starting work on the
project.  The client is capable of connecting to a Kovan server, drive a
few simple script commands, and parse and log the output.

Many script commands are unimplemented.  There are also timing issues
wherein NAND flash isn't getting read properly.  They will need to be
worked out.